Move Expiration to cw0 module and add link to source
