+++ packages/cosmwasm/src/cw3cosmwasmclient.ts
@@ -3,8 +3,8 @@ import { BroadcastMode, GasLimits, GasPrice, OfflineSigner } from "@cosmjs/launc
 
 import { CosmosMsg } from "./cosmosmsg";
 import { Account } from "./cosmwasmclient";
+import { Expiration } from "./interfaces";
 import { CosmWasmFeeTable, ExecuteResult, SigningCosmWasmClient } from "./signingcosmwasmclient";
-import { Expiration } from "./types";
 
 export enum Vote {
   Yes = "yes",
