+++ packages/cosmwasm/src/index.ts
@@ -1,4 +1,6 @@
+export { Expiration } from "./interfaces";
 export { setupWasmExtension, WasmExtension } from "./lcdapi/wasm";
+
 export { BankMsg, CosmosMsg, CustomMsg, StakingMsg, WasmMsg } from "./cosmosmsg";
 export {
   Account,
@@ -53,4 +55,3 @@ export {
   MsgUpdateAdmin,
   MsgStoreCode,
 } from "./msgs";
-export { Expiration } from "./types";
