+++ packages/cosmwasm/types/interfaces/index.d.ts
@@ -1,0 +1,1 @@
+export { Expiration } from "./cw0";
